Added Updates to the mongoConnection
