+++ comps/dbconnector/mongo/mongo_store.py
@@ -2,27 +2,8 @@
 import bson.errors as BsonError
 from bson import json_util
 import json
-from config import MONGO_HOST, MONGO_PORT, COLLECTION_NAME, DB_NAME
-import motor.motor_asyncio as motor
-from typing import Any
-
-
-class MongoClient:
-    mongo_host = MONGO_HOST
-    mongo_port = MONGO_PORT
-    conn_url = f"mongodb://{mongo_host}:{mongo_port}/"
-    db_name = DB_NAME
-
-    @staticmethod
-    def get_db_client() -> Any:
-        try:
-            client = motor.AsyncIOMotorClient(MongoClient.conn_url)
-            db = client[DB_NAME]
-            return db
-
-        except Exception as e:
-            print(e)
-            raise Exception()
+from config import COLLECTION_NAME
+from mongo_conn import MongoClient
 
 class DocumentStore:
 
@@ -41,7 +22,7 @@ class DocumentStore:
 
         try:
             inserted_conv = await self.collection.insert_one(
-                document.model_dump(by_alias=True, mode="json", exclude={"document_id"})
+                document.model_dump(by_alias=True, mode="json", exclude={"id"})
             )
             document_id = str(inserted_conv.inserted_id)
             return document_id
@@ -58,7 +39,8 @@ class DocumentStore:
                 {"$set": {"data": updated_data.model_dump(by_alias=True, mode="json"), "first_query": first_query}}
             )
             if update_result.modified_count == 1:
-                return True
+                document_id = str(update_result.inserted_id)
+                return document_id
             else:
                 return False
         except BsonError.InvalidId as e:
@@ -72,10 +54,12 @@ class DocumentStore:
         conversation_list: list = []
         try:
             cursor = self.collection.find({"user": self.user}, {"data": 0})
+            
             async for document in cursor:
+                document["id"] = str(document["_id"])
+                del document["_id"]
                 conversation_list.append(document)
-            response = json.loads(json_util.dumps(conversation_list))
-            return response
+            return conversation_list
 
         except Exception as e:
             print(e)
@@ -88,7 +72,7 @@ class DocumentStore:
             response: dict | None = await self.collection.find_one({"_id": _id, "user": self.user})
             #TODO this is a hack, need to fix this
             if response:
-                response = json.loads(json_util.dumps(response))
+                del response["_id"]
                 return response["data"]
             return None
